requirements.txt
langchain-core
langchain-community
langchain-openai
langchain-pinecone
langgraph
pydantic
python-dotenv
pandas
torch
sentence-transformers
fastapi
uvicorn
